Made BWT use less memory.

+++ GGZ/BWT.h
@@ -4,14 +4,15 @@
 #include <utility>
 #include <cstdlib>
 
+
 class Data {
 public:
-	uint8_t *data;
-	int size;
+	uint8_t * data;
+	long long size;
 
 	Data(int size) {
 		this->size = size;
-		data = (uint8_t*)malloc(size*sizeof(uint8_t));
+		data = (uint8_t*)malloc(size * sizeof(uint8_t));
 	}
 	Data() {
 		data = 0;
@@ -32,12 +33,13 @@ public:
 	}
 };
 
+
 class BWT {
 public:
 	//static pair<string, int> Transform(string toTransform);
 	//static std::pair<Data, short> Transform(uint8_t* toTransform, int size);
-	static Data Reverse(Data toReverse, int index);
-	static void Reverse(Data *toReverse, unsigned short index);
-	static unsigned short Transform(Data* toTransform);
+	static Data Reverse(Data toReverse, long long index);
+	static void Reverse(Data *toReverse, long long index);
+	static long long Transform(Data* toTransform);
 
 };
